requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
boto3==1.35.40
botocore==1.35.40
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==5.1.2
django-cors-headers==4.5.0
django-filter==24.3
django-phonenumber-field==8.0.0
django-ses==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
idna==3.10
jmespath==1.0.1
kombu==5.4.2
oauthlib==3.2.2
paypalrestsdk==1.13.3
phonenumbers==8.13.47
prompt_toolkit==3.0.48
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.3
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
